third_party/blink/web_tests/fast/dom/Window/window-open-pending-url.html

<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.setPopupBlockingEnabled(false);
    testRunner.waitUntilDone();
}

var popup = window.open('resources/destination.html');
document.write('location: ' + popup.location + '<br/>');
var props = ['href', 'pathname', 'host', 'hostname', 'hash', 'port',
             'protocol', 'search'];
for (var i = 0; i < props.length; ++i) {
  var prop = props[i];
  document.write(prop + ': ' + popup.location[prop] + '<br/>');
}
</script>
