third_party/WebKit/LayoutTests/fast/dom/Window/window-open-pending-url-expected.txt

location: about:blank
href: about:blank
pathname: blank
host:
hostname:
hash:
port:
protocol: about:
search:
